Move Vite entry to src/main
