+++ raikomix-youtube-dj_1.0/vercel.json
@@ -1,0 +1,5 @@
+{
+  "framework": "vite",
+  "buildCommand": "npm run build",
+  "outputDirectory": "dist"
+}
